OvmfPkg/AcpiPlatformDxe: patch FADT PSCI bits if FDT advertises it
EL3 firmware may implement PSCI interface on aarch64 platforms, including qemu-tcg-aarch64. However, EL3 firmware does not usually own pulling and deploying ACPI tables from qemu fw_cfg. Thus the only way EL3 can advertise PSCI on qemu is in FDT. One such EL3 fw is ARM trusted firmware. Qemu itself also won't advertise PSCI in either FDT or ACPI if EL3 firmware is present.

PSCI can be advertised in both FDT and ACPI, and Hyper-V/NT kernel expect to see all information published in ACPI. To better support running Hyper-V/NT on qemu-tcg-aarch64 with EDK2 as UEFI implementation and ARM trusted firmware as EL3 PSCI implementation we can patch in PSCI bits in ACPI FADT when pulling tables from fw_cfg if PSCI node is advertised in FDT. EDK2 owns ACPI table publishing and is also aware of FDT on arm, so it is ideally poised to handle this.

This change illustrates how it could potentially be done. I am looking for comments on overall validity of the idea to patch FADT and whether or not this particular approach of handling it in AcpiPlatformDxe is the way to do it.

Signed-off-by: Evgeny Iakovlev <eiakovlev@linux.microsoft.com>